refactor: add navigation module
